# Skip asset pipeline resolution for for Service nodes in TreeNodeBuilder#node_icon