@Fryguy 2016-09-01
@NickLaMuro I find this interesting, because it *shouldn't* be needed, but perhaps I'm missing something?  Are you doing this in production mode or development mode?  In production mode are you using precompiled assets (which is what is actually done)?

In development mode, you don't precompile the assets so the lookup is dynamic.  That is intentional, as in dev mode you also have things like code-reloading, so although it's inherently slower, it's more flexible.  So, if someone modifies the image on the fly, it's *supposed* to change.  I don't think we should add caching just for dev mode.

In production mode, the assets are precompiled, and there is a manifest.  When the `image_path` helper runs, it does a lookup into this manifest to determine the precompiled image location, which *should* be like a Hash lookup and thus O(1).  If it's not, I'd be really surprised, and this kind of cache should be put into sprockets or perhaps Rails itself.