@NickLaMuro 2016-09-01
@Fryguy Yeah, I was wondering the same thing myself right after I closed my laptop, and the first thing I am going to do is double check this in production and such.  I had done tests in prod with similar results (but might have had a dated `precompile` or something), and those metrics are from a `dev` mode, so if this is a bunch of malarky I am spewing at 8pm on a Wednesday, then this should be closed immediately...

At the very least, I want to test this on an appliance with a before and after as well (if my tests remain consistent in production mode).

And yes, I only want this if it is a prod fix.  If there is a sprockets way of doing it (assuming it for some reason isn't being intentionally being ignored), I would prefer that over this hackjob.