@NickLaMuro 2016-09-01
For the record, logged into an appliance that @dmetzger57 had setup and checked for a `public/assets/manifest.json` (default location if I am not mistaken?), as well as running a `find . -name manifest` from the app directory and found nothing, same locally (even after booting up the server).

I will look to see if there is a config we are missing as well, and if this is not there, then I suspect that explains my results (also, we should probably just close this PR... it was dumb from the start...).