@NickLaMuro 2016-09-12
Re-opening because I have tested this out on `/service/explorer` and found that it gave a 4-5 second boost when present (with some updates based on code review).  Some further investigation into "why" needs to be done though.