@NickLaMuro 2016-09-12
@Fryguy @chessbyte I have re-opened this PR since I have found a performance improvement use case in `/service/explorer` that will be present in Production mode.

I have update the commit/commit message to address your issues and also explain the use case.  In short, we are using customer images from the DB when rendering trees like the `/service/explorer`.  The cached hash ends up looking something like this:

```ruby
{
  "../../../pictures/XXXr1.jpg"=>"/images/100/../../../pictures/XXXr1.jpg",
  "../../../pictures/XXXr2.jpg"=>"/images/100/../../../pictures/XXXr2.jpg",
  "../../../pictures/XXXr3.jpg"=>"/images/100/../../../pictures/XXXr4.jpg"
}
```

This means that these will **not** end up on the file digest of sprockets, and thus cause some slow down.  Not sure if there is a better place to add this that is more inclusive to this, so I am open to suggestions on a better place for this cache live, but it cuts the render time (after Keenan's patches) for `/service/explorer` in half on large trees.

/cc @kbrock