@kbrock 2016-09-12
:+1: this cuts down time singificantly