@martinpovolny 2016-09-13
Sure, we can merge this as it helps us.

But this seems to be a Rails issue. We should not need to cache `ActionController::Base.helpers.image_path`, shoud we?

If it's really helpful we should cache all these calls over all of ManageIQ.

ping @matthewd : please, take a look