@kbrock 2016-09-13
And @matthewd is aware of this problem in particular. I think most of the time is spent including modules or something. A helper class that caches a reference to `ActionController::Base.helpers` may suffice.