@matthewd 2016-09-13
> `"../../../pictures/XXXr1.jpg"`

So should this just not be using `image_path` at all for such paths?

If the stored value was instead `/pictures/XXXr1.jpg`, then the method could be more like:

```ruby
def node_icon(icon)
  if icon.start_with?('/')
    icon
  else
    ActionController::Base.helpers.image_path("100/#{icon}")
  end
end
```

and the slow / ill-fated asset lookup would never occur. (There are obviously existing-data-compatible variants of the above, looking for `../`, if that's more convenient.)

We can probably do something about the fact we try harder than we need to before we give up, but an asset path that starts inside a valid subdir then works its way out of the whole asset system seems.. unusual.

"Cache the slow pointless thing" is an improvement, but avoiding it seems even better :wink:


---

My specific recollection was that it may have become necessary to cache `ActionController::Base.helpers` -- that it used to be internally memoized for other reasons, but no longer. But that would be "re-defining a bunch of helper methods" slow, not "scanning filesystem" slow.