@kbrock 2016-09-13
@matthewd heh. I was just coding this up

```ruby
class ImageHelper
  extend Forwardable

  def self.helpers
    @@helpers = ActionController::Base.helpers # evil cache
  end
  self << class
  def_delegators :helpers, :image_path
  end
end
```