@NickLaMuro 2016-09-13
@matthewd @kbrock Thanks, yeah, I think that is the smarter approach, and I think I will go with the simplified version of that in this case (sans helper for now).  Thanks for the help all!