@chessbyte 2016-09-15
@NickLaMuro You may want to update the PR title and description with the new approach.