@chessbyte 2016-09-30
This was on master before Euwe branch was created.