@chessbyte 2016-10-07
@dmetzger57 @simaishi need Darga BZ for backport.