@simaishi 2016-10-10
Darga BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1383469