@chessbyte 2016-10-11
Darga Backport details:
```sh
$ git log
commit 7feca7f49108cb91f82ff493229c59ad6abf2218
Author: Martin Povolny <mpovolny@redhat.com>
Date:   Wed Sep 21 18:10:09 2016 +0200

    Merge pull request #10929 from NickLaMuro/faster-assets-on-tree-node-builder
    
    Skip asset pipeline resolution for for Service nodes in TreeNodeBuilder#node_icon
    (cherry picked from commit 51af6d736c3d3fee5e284ef89a359feaf990b788)
    
    https://bugzilla.redhat.com/show_bug.cgi?id=1383469
```